Refactored repo to better structure
