+++ go.mod
@@ -1,0 +1,5 @@
+module github.com/wizedkyle/sumocli
+
+go 1.15
+
+require github.com/spf13/cobra v1.1.1
