--- main.go
@@ -1,9 +1,0 @@
-package main
-
-import (
-	"github.com/wizedkyle/sumocli/cmd"
-)
-
-func main() {
-	cmd.Execute()
-}
