+++ pkg/cmd/users/users.go
@@ -1,8 +1,9 @@
-package cmd
+package users
 
 import (
 	"github.com/spf13/cobra"
-	"github.com/wizedkyle/sumocli/users"
+	"github.com/wizedkyle/sumocli/cmd"
+	"github.com/wizedkyle/sumocli/pkg/cmd/users/create"
 )
 
 var (
@@ -17,13 +18,13 @@ var usersCreateCmd = &cobra.Command{
 	Short: "Creates a Sumo Logic user",
 	Long:  `Creates a Sumo Logic user by specifying the first name, last name, email and roleIds`,
 	Run: func(cmd *cobra.Command, args []string) {
-		users.CreateUser(firstName, lastName, emailAddress, roleIds)
+		create.CreateUser(firstName, lastName, emailAddress, roleIds)
 
 	},
 }
 
 func init() {
-	createCmd.AddCommand(usersCreateCmd)
+	cmd.createCmd.AddCommand(usersCreateCmd)
 
 	usersCreateCmd.PersistentFlags().StringVar(&firstName, "fn", "", "First name of the user")
 	usersCreateCmd.PersistentFlags().StringVar(&lastName, "ln", "", "Last name of the user")
