+++ pkg/cmdutil/log.go
@@ -1,4 +1,4 @@
-package util
+package cmdutil
 
 import "log"
