+++ pkg/cmdutil/output.go
@@ -1,4 +1,4 @@
-package util
+package cmdutil
 
 import (
 	"fmt"
